ops: migration script to generate hashes for legacy submissions
